pyproject.toml
[project]
name = "autocomp"
version = "0.1.1"
authors = [
    { name="charleshong3", email="charleshong@berkeley.edu" },
]
description = "Automation compilation for diverse hardware backends ."
readme = "README.md"
license = { file="LICENSE" }
requires-python = ">=3.10.0"
dependencies = [
    "openai",
    "google-genai",
    "anthropic",
    "boto3",
    "numpy",
    "wandb",
    "together",
    "pyyaml",
]

[project.optional-dependencies]
agent-builder = [
    "pymupdf",
    "requests",
    "beautifulsoup4",
    "html2text",
]
dev = [
    "pytest",
]

[project.urls]
"Homepage" = "https://github.com/ucb-bar/autocomp"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["autocomp", "autocomp.*"]

[tool.pytest.ini_options]
testpaths = ["autocomp_tests"]
